+++ addie/plot/__init__.py
@@ -1,0 +1,5 @@
+from __future__ import absolute_import
+
+from .canvas import Qt4MplCanvas  # noqa
+from .navigationtoolbar import MyNavigationToolbar  # noqa
+from .indicatormanager import IndicatorManager  # noqa
