+++ addie/utilities/mplgraphicsview.py
@@ -3,285 +3,10 @@ from __future__ import (absolute_import, division, print_function)
 import os
 import numpy as np
 
-from qtpy import PYQT4, PYQT5
-from qtpy.QtCore import (Signal)
-from qtpy.QtWidgets import (QSizePolicy, QVBoxLayout, QWidget)
+from qtpy.QtWidgets import QVBoxLayout, QWidget
 
-if PYQT5:
-    from matplotlib.backends.backend_qt5agg import FigureCanvasQTAgg as FigureCanvas
-    from matplotlib.backends.backend_qt5agg import NavigationToolbar2QT as NavigationToolbar2
-elif PYQT4:
-    from matplotlib.backends.backend_qt4agg import FigureCanvasQTAgg as FigureCanvas
-    from matplotlib.backends.backend_qt4agg import NavigationToolbar2QT as NavigationToolbar2
-else:
-    raise ImportError('do not know which matplotlib backend to use')
-
-import matplotlib.image  # noqa
-from matplotlib.figure import Figure  # noqa
-
-MplLineStyles = ['-', '--', '-.', ':', 'None', ' ', '']
-MplLineMarkers = [
-    ". (point         )",
-    "* (star          )",
-    "x (x             )",
-    "o (circle        )",
-    "s (square        )",
-    "D (diamond       )",
-    ", (pixel         )",
-    "v (triangle_down )",
-    "^ (triangle_up   )",
-    "< (triangle_left )",
-    "> (triangle_right)",
-    "1 (tri_down      )",
-    "2 (tri_up        )",
-    "3 (tri_left      )",
-    "4 (tri_right     )",
-    "8 (octagon       )",
-    "p (pentagon      )",
-    "h (hexagon1      )",
-    "H (hexagon2      )",
-    "+ (plus          )",
-    "d (thin_diamond  )",
-    "| (vline         )",
-    "_ (hline         )",
-    "None (nothing    )"]
-
-# Note: in colors, "white" is removed
-MplBasicColors = [
-    "black",
-    "red",
-    "blue",
-    "green",
-    "cyan",
-    "magenta",
-    "yellow"]
-
-
-class IndicatorManager(object):
-    """ Manager for all indicator lines
-
-    Indicator's Type =
-    - 0: horizontal.  moving along Y-direction. [x_min, x_max], [y, y];
-    - 1: vertical. moving along X-direction. [x, x], [y_min, y_max];
-    - 2: 2-way. moving in any direction. [x_min, x_max], [y, y], [x, x], [y_min, y_max].
-    """
-
-    def __init__(self):
-        # Auto color index
-        self._colorIndex = 0
-        # Auto line ID
-        self._autoLineID = 1
-
-        self._lineManager = dict()
-        self._canvasLineKeyDict = dict()
-        self._indicatorTypeDict = dict()  # value: 0 (horizontal), 1 (vertical), 2 (2-way)
-
-    def add_2way_indicator(self, x, x_min, x_max, y, y_min, y_max, color):
-        # Set up indicator ID
-        this_id = str(self._autoLineID)
-        self._autoLineID += 1
-
-        # Set up vectors
-        vec_x_horizontal = np.array([x_min, x_max])
-        vec_y_horizontal = np.array([y, y])
-
-        vec_x_vertical = np.array([x, x])
-        vec_y_vertical = np.array([y_min, y_max])
-
-        #
-        self._lineManager[this_id] = [vec_x_horizontal, vec_y_horizontal, vec_x_vertical, vec_y_vertical, color]
-        self._indicatorTypeDict[this_id] = 2
-
-        return this_id
-
-    def add_horizontal_indicator(self, y, x_min, x_max, color):
-        """
-        Add a horizontal indicator moving vertically
-        """
-        # Get ID
-        this_id = str(self._autoLineID)
-        self._autoLineID += 1
-
-        #
-        vec_x = np.array([x_min, x_max])
-        vec_y = np.array([y, y])
-
-        #
-        self._lineManager[this_id] = [vec_x, vec_y, color]
-        self._indicatorTypeDict[this_id] = 0
-
-        return this_id
-
-    def add_vertical_indicator(self, x, y_min, y_max, color):
-        """
-        Add a vertical indicator to data structure moving horizontally
-        :return: indicator ID as an integer
-        """
-        # Get ID
-        this_id = self._autoLineID
-        self._autoLineID += 1
-
-        # form vec x and vec y
-        vec_x = np.array([x, x])
-        vec_y = np.array([y_min, y_max])
-
-        #
-        self._lineManager[this_id] = [vec_x, vec_y, color]
-        self._indicatorTypeDict[this_id] = 1
-
-        return this_id
-
-    def delete(self, indicator_id):
-        """
-        Delete indicator
-        """
-        del self._lineManager[indicator_id]
-        del self._canvasLineKeyDict[indicator_id]
-        del self._indicatorTypeDict[indicator_id]
-
-    def get_canvas_line_index(self, indicator_id):
-        """
-        Get a line's ID (on canvas) from an indicator ID
-        """
-        assert isinstance(indicator_id, int)
-
-        if indicator_id not in self._canvasLineKeyDict:
-            raise RuntimeError('Indicator ID %s cannot be found. Current keys are %s.' % (
-                indicator_id, str(sorted(self._canvasLineKeyDict.keys()))
-            ))
-        return self._canvasLineKeyDict[indicator_id]
-
-    def get_line_type(self, my_id):
-        return self._indicatorTypeDict[my_id]
-
-    def get_2way_data(self, line_id):
-        assert line_id in self._indicatorTypeDict, 'blabla'
-        assert self._indicatorTypeDict[line_id] == 2, 'blabla'
-
-        vec_set = [self._lineManager[line_id][0:2], self._lineManager[line_id][2:4]]
-
-        return vec_set
-
-    def get_data(self, line_id):
-        """
-        Get line's vector x and vector y
-        :return: 2-tuple of numpy arrays
-        """
-        return self._lineManager[line_id][0], self._lineManager[line_id][1]
-
-    def get_indicator_key(self, x, y):
-        """ Get indicator's key with position
-        """
-        if x is None and y is None:
-            raise RuntimeError('It is not allowed to have both X and Y are none to get indicator key.')
-
-        ret_key = None
-
-        for line_key in self._lineManager:
-
-            if x is not None and y is not None:
-                # 2 way
-                raise NotImplementedError('ASAP')
-            elif x is not None and self._indicatorTypeDict[line_key] == 1:
-                # vertical indicator moving along X
-                if abs(self._lineManager[line_key][0][0] - x) < 1.0E-2:
-                    return line_key
-            elif y is not None and self._indicatorTypeDict[line_key] == 0:
-                # horizontal indicator moving along Y
-                if abs(self._lineManager[line_key][1][0] - y) < 1.0E-2:
-                    return line_key
-        # END-FOR
-
-        return ret_key
-
-    @staticmethod
-    def get_line_style(line_id=None):
-        if line_id is not None:
-            style = '--'
-        else:
-            style = '--'
-
-        return style
-
-    def get_live_indicator_ids(self):
-        return sorted(self._lineManager.keys())
-
-    @staticmethod
-    def get_marker():
-        """
-        Get the marker a line
-        """
-        return '.'
-
-    def get_next_color(self):
-        """
-        Get next color by auto color index
-        :return: string as color
-        """
-        next_color = MplBasicColors[self._colorIndex]
-
-        # Advance and possibly reset color scheme
-        self._colorIndex += 1
-        if self._colorIndex == len(MplBasicColors):
-            self._colorIndex = 0
-
-        return next_color
-
-    def set_canvas_line_index(self, my_id, canvas_line_index):
-        self._canvasLineKeyDict[my_id] = canvas_line_index
-
-    def set_position(self, my_id, pos_x, pos_y):
-        """ Set the indicator to a new position
-        """
-        if self._indicatorTypeDict[my_id] == 0:
-            # horizontal
-            self._lineManager[my_id][1][0] = pos_y
-            self._lineManager[my_id][1][1] = pos_y
-
-        elif self._indicatorTypeDict[my_id] == 1:
-            # vertical
-            self._lineManager[my_id][0][0] = pos_x
-            self._lineManager[my_id][0][1] = pos_x
-
-        elif self._indicatorTypeDict[my_id] == 2:
-            # 2-way
-            self._lineManager[my_id][0] = pos_x
-            self._lineManager[my_id][1] = pos_y
-
-        else:
-            raise RuntimeError('Unsupported indicator of type %d' % self._indicatorTypeDict[my_id])
-
-        self._lineManager[my_id][2] = 'black'
-
-    def shift(self, my_id, dx, dy):
-        if self._indicatorTypeDict[my_id] == 0:
-            # horizontal
-            self._lineManager[my_id][1] += dy
-
-        elif self._indicatorTypeDict[my_id] == 1:
-            # vertical
-            self._lineManager[my_id][0] += dx
-
-        elif self._indicatorTypeDict[my_id] == 2:
-            # 2-way
-            self._lineManager[my_id][2] += dx
-            self._lineManager[my_id][1] += dy
-
-        else:
-            raise RuntimeError('Unsupported indicator of type %d' % self._indicatorTypeDict[my_id])
-
-    def update_indicators_range(self, x_range, y_range):
-        """
-        Update indicator's range
-        """
-        for i_id in self._lineManager:
-            # NEXT - Need a new flag for direction of the indicating line, vertical or horizontal
-            if True:
-                self._lineManager[i_id][1][0] = y_range[0]
-                self._lineManager[i_id][1][-1] = y_range[1]
-            else:
-                self._lineManager[i_id][0][0] = x_range[0]
-                self._lineManager[i_id][0][-1] = x_range[1]
+from addie.plot import IndicatorManager, MyNavigationToolbar, Qt4MplCanvas
+from addie.plot.constants import MplBasicColors, MplLineMarkers, MplLineStyles
 
 
 class MplGraphicsView(QWidget):
@@ -858,703 +583,3 @@ class MplGraphicsView(QWidget):
         """
         """
         self._myLineMarkerColorIndex = newindex
-
-
-class Qt4MplCanvas(FigureCanvas):
-    """  A customized Qt widget for matplotlib figure.
-    It can be used to replace GraphicsView of QtGui
-    """
-
-    def __init__(self, parent):
-        # from mpl_toolkits.axes_grid1 import host_subplot
-        # import mpl_toolkits.axisartist as AA
-        # import matplotlib.pyplot as plt
-
-        # Instantiating matplotlib Figure
-        self.fig = Figure()
-        self.fig.patch.set_facecolor('white')
-
-        if True:
-            self.axes = self.fig.add_subplot(111, projection='mantid')  # return: matplotlib.axes.AxesSubplot
-            self.fig.subplots_adjust(bottom=0.15)
-            self.axes2 = None
-        else:
-            self.axes = self.fig.add_host_subplot(111, projection='mantid')
-
-        # Initialize parent class and set parent
-        FigureCanvas.__init__(self, self.fig)
-        self.setParent(parent)
-
-        # Set size policy to be able to expanding and resizable with frame
-        FigureCanvas.setSizePolicy(self, QSizePolicy.Expanding, QSizePolicy.Expanding)
-        FigureCanvas.updateGeometry(self)
-
-        # Variables to manage all lines/subplot
-        self._lineDict = {}
-        self._lineIndex = 0
-
-        # legend and color bar
-        self._colorBar = None
-        self._isLegendOn = False
-        self._legendFontSize = 8
-
-    @property
-    def is_legend_on(self):
-        """
-        check whether the legend is shown or hide
-        """
-        return self._isLegendOn
-
-    def add_arrow(self, start_x, start_y, stop_x, stop_y):
-        """
-        0, 0, 0.5, 0.5, head_width=0.05, head_length=0.1, fc='k', ec='k')
-        :return:
-        """
-        head_width = 0.05
-        head_length = 0.1
-        fc = 'k'
-        ec = 'k'
-
-        self.axes.arrrow(start_x, start_y, stop_x, stop_y, head_width,
-                         head_length, fc, ec)
-
-    def add_plot_1d(self, vec_x, vec_y, y_err=None, color=None, label="", x_label=None, y_label=None,
-                    marker=None, line_style=None, line_width=1, alpha=1., show_legend=True):
-        """
-        :param vec_x: numpy array X
-        :param vec_y: numpy array Y
-        :param y_err:
-        :param color:
-        :param label:
-        :param x_label:
-        :param y_label:
-        :param marker:
-        :param line_style:
-        :param line_width:
-        :param alpha:
-        :param show_legend:
-        :return: new key
-        """
-        # Check input
-        if isinstance(vec_x, np.ndarray) is False or isinstance(vec_y, np.ndarray) is False:
-            raise NotImplementedError('Input vec_x or vec_y for addPlot() must be numpy.array.')
-        plot_error = y_err is not None
-        if plot_error is True:
-            if isinstance(y_err, np.ndarray) is False:
-                raise NotImplementedError('Input y_err must be either None or numpy.array.')
-
-        if len(vec_x) != len(vec_y):
-            raise NotImplementedError('Input vec_x and vec_y must have same size.')
-        if plot_error is True and len(y_err) != len(vec_x):
-            raise NotImplementedError('Input vec_x, vec_y and y_error must have same size.')
-
-        # Hold previous data
-        self.axes.hold(True)
-
-        # set x-axis and y-axis label
-        if x_label is not None:
-            self.axes.set_xlabel(x_label, fontsize=16)
-        if y_label is not None:
-            self.axes.set_ylabel(y_label, fontsize=16)
-
-        # process inputs and defaults
-        if color is None:
-            color = (0, 1, 0, 1)
-        if marker is None:
-            marker = 'None'
-        if line_style is None:
-            line_style = '-'
-
-        # color must be RGBA (4-tuple)
-        if plot_error is False:
-            # return: list of matplotlib.lines.Line2D object
-            r = self.axes.plot(vec_x, vec_y, color=color, marker=marker, markersize=2, linestyle=line_style,
-                               label=label, linewidth=line_width, alpha=alpha,)
-        else:
-            r = self.axes.errorbar(vec_x, vec_y, yerr=y_err, color=color, marker=marker, linestyle=line_style,
-                                   label=label, linewidth=line_width, alpha=alpha,
-                                   markersize=40)
-
-        self.axes.set_aspect('auto')
-
-        # set x-axis and y-axis label
-        if x_label is not None:
-            self.axes.set_xlabel(x_label, fontsize=20)
-        if y_label is not None:
-            self.axes.set_ylabel(y_label, fontsize=20)
-
-        # set/update legend
-        if show_legend:
-            self._setup_legend()
-
-        # Register
-        line_key = self._lineIndex
-        if len(r) == 1:
-            self._lineDict[line_key] = r[0]
-            self._lineIndex += 1
-        else:
-            msg = 'Return from plot is a %d-tuple: %s.. \n' % (len(r), r)
-            for i_r in range(len(r)):
-                msg += 'r[%d] = %s\n' % (i_r, str(r[i_r]))
-            raise NotImplementedError(msg)
-
-        # Flush/commit
-        self.draw()
-
-        return line_key
-
-    def add_1d_plot_right(self, x, y, color=None, label="", x_label=None, ylabel=None, marker=None, linestyle=None,
-                          linewidth=1):
-        """ Add a line (1-d plot) at right axis
-        """
-        if self.axes2 is None:
-            self.axes2 = self.axes.twinx()
-
-        # Hold previous data
-        self.axes2.hold(True)
-
-        # Default
-        if color is None:
-            color = (0, 1, 0, 1)
-        if marker is None:
-            marker = 'o'
-        if linestyle is None:
-            linestyle = '-'
-
-        # Special default
-        if len(label) == 0:
-            label = 'right'
-            color = 'red'
-
-        # color must be RGBA (4-tuple)
-        r = self.axes2.plot(x, y, color=color, marker=marker, linestyle=linestyle,
-                            label=label, linewidth=linewidth)
-        # return: list of matplotlib.lines.Line2D object
-
-        self.axes2.set_aspect('auto')
-
-        # set x-axis and y-axis label
-        if x_label is not None:
-            self.axes2.set_xlabel(x_label, fontsize=20)
-        if ylabel is not None:
-            self.axes2.set_ylabel(ylabel, fontsize=20)
-
-        # set/update legend
-        self._setup_legend()
-
-        # Register
-        line_key = -1
-        if len(r) == 1:
-            line_key = self._lineIndex
-            self._lineDict[line_key] = r[0]
-            self._lineIndex += 1
-        else:
-            print("Impoooooooooooooooosible!")
-
-        # Flush/commit
-        self.draw()
-
-        return line_key
-
-    def addPlot2D(self, array2d, xmin, xmax, ymin, ymax, holdprev, yticklabels=None):
-        """ Add a 2D plot
-
-        Arguments:
-         - yticklabels :: list of string for y ticks
-        """
-        # Release the current image
-        self.axes.hold(holdprev)
-
-        # Do plot
-        # y ticks will be shown on line 1, 4, 23, 24 and 30
-        # yticks = [1, 4, 23, 24, 30]
-        # self.axes.set_yticks(yticks)
-
-        # show image
-        imgplot = self.axes.imshow(array2d, extent=[xmin, xmax, ymin, ymax], interpolation='none')
-        # set y ticks as an option:
-        if yticklabels is not None:
-            # it will always label the first N ticks even image is zoomed in
-            print("--------> [FixMe]: The way to set up the Y-axis ticks is wrong!")
-            # self.axes.set_yticklabels(yticklabels)
-
-        # explicitly set aspect ratio of the image
-        self.axes.set_aspect('auto')
-
-        # Set color bar.  plt.colorbar() does not work!
-        if self._colorBar is None:
-            # set color map type
-            imgplot.set_cmap('spectral')
-            self._colorBar = self.fig.colorbar(imgplot)
-        else:
-            self._colorBar.update_bruteforce(imgplot)
-
-        # Flush...
-        self._flush()
-
-    def add_contour_plot(self, vec_x, vec_y, matrix_z):
-        # create mesh grid
-        grid_x, grid_y = np.meshgrid(vec_x, vec_y)
-
-        # check size
-        assert grid_x.shape == matrix_z.shape, 'Size of X (%d) and Y (%d) must match size of Z (%s).' \
-                                               '' % (len(vec_x), len(vec_y), matrix_z.shape)
-
-        # Release the current image
-        self.axes.hold(False)
-
-        # Do plot
-        contour_plot = self.axes.contourf(grid_x, grid_y, matrix_z, 100)
-
-        labels = [item.get_text() for item in self.axes.get_yticklabels()]
-        print('[DB...BAT] Number of Y labels = ', len(labels), ', Number of Y = ', len(vec_y))
-
-        # TODO/ISSUE/55: how to make this part more powerful
-        if len(labels) == 2*len(vec_y) - 1:
-            new_labels = [''] * len(labels)
-            for i in range(len(vec_y)):
-                new_labels[i*2] = '%d' % int(vec_y[i])
-            self.axes.set_yticklabels(new_labels)
-
-        # explicitly set aspect ratio of the image
-        self.axes.set_aspect('auto')
-
-        # Set color bar.  plt.colorbar() does not work!
-        if self._colorBar is None:
-            # set color map type
-            contour_plot.set_cmap('spectral')
-            self._colorBar = self.fig.colorbar(contour_plot)
-        else:
-            self._colorBar.update_bruteforce(contour_plot)
-
-        # Flush...
-        self._flush()
-
-    def addImage(self, imagefilename):
-        """ Add an image by file
-        """
-        #import matplotlib.image as mpimg
-
-        # set aspect to auto mode
-        self.axes.set_aspect('auto')
-
-        img = matplotlib.image.imread(str(imagefilename))
-        # lum_img = img[:,:,0]
-        # FUTURE : refactor for image size, interpolation and origin
-        imgplot = self.axes.imshow(img, extent=[0, 1000, 800, 0], interpolation='none', origin='lower')
-
-        # Set color bar.  plt.colorbar() does not work!
-        if self._colorBar is None:
-            # set color map type
-            imgplot.set_cmap('spectral')
-            self._colorBar = self.fig.colorbar(imgplot)
-        else:
-            self._colorBar.update_bruteforce(imgplot)
-
-        self._flush()
-
-    def clear_all_1d_plots(self):
-        """ Remove all lines from the canvas
-        """
-        for ikey in list(self._lineDict.keys()):
-            plot = self._lineDict[ikey]
-            if plot is None:
-                continue
-            if isinstance(plot, tuple) is False:
-                try:
-                    self.axes.lines.remove(plot)
-                except ValueError as e:
-                    print("[Error] Plot %s is not in axes.lines which has %d lines. Error mesage: %s" % (
-                        str(plot), len(self.axes.lines), str(e)))
-                del self._lineDict[ikey]
-            else:
-                # error bar
-                plot[0].remove()
-                for line in plot[1]:
-                    line.remove()
-                for line in plot[2]:
-                    line.remove()
-                del self._lineDict[ikey]
-            # ENDIF(plot)
-        # ENDFOR
-
-        self._setup_legend()
-
-        self.draw()
-
-    def clear_canvas(self):
-        """ Clear data including lines and image from canvas
-        """
-        # clear the image for next operation
-        self.axes.hold(False)
-
-        # Clear all lines
-        self.clear_all_1d_plots()
-
-        # clear image
-        self.axes.cla()
-        # Try to clear the color bar
-        if len(self.fig.axes) > 1:
-            self.fig.delaxes(self.fig.axes[1])
-            self._colorBar = None
-            # This clears the space claimed by color bar but destroys sub_plot too.
-            self.fig.clear()
-            # Re-create subplot
-            self.axes = self.fig.add_subplot(111)
-            self.fig.subplots_adjust(bottom=0.15)
-
-        # flush/commit
-        self._flush()
-
-    def decrease_legend_font_size(self):
-        """
-        reset the legend with the new font size
-        Returns:
-
-        """
-        # minimum legend font size is 2! return if it already uses the smallest font size.
-        if self._legendFontSize <= 2:
-            return
-
-        self._legendFontSize -= 1
-        self._setup_legend(font_size=self._legendFontSize)
-
-        self.draw()
-
-    def getLastPlotIndexKey(self):
-        """ Get the index/key of the last added line
-        """
-        return self._lineIndex-1
-
-    def getPlot(self):
-        """ reture figure's axes to expose the matplotlib figure to PyQt client
-        """
-        return self.axes
-
-    def getXLimit(self):
-        """ Get limit of Y-axis
-        """
-        return self.axes.get_xlim()
-
-    def getYLimit(self):
-        """ Get limit of Y-axis
-        """
-        return self.axes.get_ylim()
-
-    def hide_legend(self):
-        """
-        hide the legend if it is not None
-        """
-        if self.axes.legend() is not None:
-            # set visible to be False and re-draw
-            self.axes.legend().set_visible(False)
-            self.draw()
-
-        self._isLegendOn = False
-
-    def increase_legend_font_size(self):
-        """
-        reset the legend with the new font size
-        """
-        self._legendFontSize += 1
-
-        self._setup_legend(font_size=self._legendFontSize)
-
-        self.draw()
-
-    def setXYLimit(self, xmin, xmax, ymin, ymax):
-        # for X
-        xlims = self.axes.get_xlim()
-        xlims = list(xlims)
-        if xmin is not None:
-            xlims[0] = xmin
-        if xmax is not None:
-            xlims[1] = xmax
-        self.axes.set_xlim(xlims)
-
-        # for Y
-        ylims = self.axes.get_ylim()
-        ylims = list(ylims)
-        if ymin is not None:
-            ylims[0] = ymin
-        if ymax is not None:
-            ylims[1] = ymax
-        self.axes.set_ylim(ylims)
-
-        # try draw
-        self.draw()
-
-    def set_title(self, title, color):
-        # TODO/NOW - doc & etc
-
-        self.axes.set_title(title, loc='center', color=color)
-
-        self.draw()
-
-    def remove_plot_1d(self, plot_key):
-        """ Remove the line with its index as key
-        """
-        # Get all lines in list
-        lines = self.axes.lines
-        assert isinstance(lines, list), 'Lines must be list'
-
-        if plot_key in self._lineDict:
-            try:
-                self.axes.lines.remove(self._lineDict[plot_key])
-            except ValueError as r_error:
-                error_message = 'Unable to remove to 1D line %s (ID=%d) due to %s.' % (str(self._lineDict[plot_key]),
-                                                                                       plot_key, str(r_error))
-                raise RuntimeError(error_message)
-            # remove the plot key from dictionary
-            del self._lineDict[plot_key]
-        else:
-            raise RuntimeError('Line with ID %s is not recorded.' % plot_key)
-
-        self._setup_legend(location='best', font_size=self._legendFontSize)
-
-        # Draw
-        self.draw()
-
-    def show_legend(self):
-        """
-        show the legend if the legend is not None
-        Returns:
-
-        """
-        if self.axes.legend() is not None:
-            # set visible to be True and re-draw
-            # self.axes.legend().set_visible(True)
-            self._setup_legend(font_size=self._legendFontSize)
-            self.draw()
-
-            # set flag on
-            self._isLegendOn = True
-
-    def updateLine(self, ikey, vecx=None, vecy=None, linestyle=None, linecolor=None, marker=None, markercolor=None):
-        """
-        Update a plot line or a series plot line
-        """
-        line = self._lineDict[ikey]
-        if line is None:
-            print('[ERROR] Line (key = %d) is None. Unable to update' % ikey)
-            return
-
-        if vecx is not None and vecy is not None:
-            line.set_xdata(vecx)
-            line.set_ydata(vecy)
-
-        if linecolor is not None:
-            line.set_color(linecolor)
-
-        if linestyle is not None:
-            line.set_linestyle(linestyle)
-
-        if marker is not None:
-            line.set_marker(marker)
-
-        if markercolor is not None:
-            line.set_markerfacecolor(markercolor)
-
-        oldlabel = line.get_label()
-        line.set_label(oldlabel)
-
-        self._setup_legend()
-
-        # commit
-        self.draw()
-
-    def get_data(self, line_id):
-        """
-        Get vecX and vecY from line object in matplotlib
-        :param line_id:
-        :return: 2-tuple as vector X and vector Y
-        """
-        # check
-        if line_id not in self._lineDict:
-            raise KeyError('Line ID %s does not exist.' % str(line_id))
-
-        # get line
-        line = self._lineDict[line_id]
-        if line is None:
-            raise RuntimeError('Line ID %s has been removed.' % line_id)
-
-        return line.get_xdata(), line.get_ydata()
-
-    def getLineStyleList(self):
-        return MplLineStyles
-
-    def getLineMarkerList(self):
-        return MplLineMarkers
-
-    def getLineBasicColorList(self):
-        return MplBasicColors
-
-    def getDefaultColorMarkerComboList(self):
-        """ Get a list of line/marker color and marker style combination
-        as default to add more and more line to plot
-        """
-        combo_list = list()
-        num_markers = len(MplLineMarkers)
-        num_colors = len(MplBasicColors)
-
-        for i in range(num_markers):
-            marker = MplLineMarkers[i]
-            for j in range(num_colors):
-                color = MplBasicColors[j]
-                combo_list.append((marker, color))
-            # ENDFOR (j)
-        # ENDFOR(i)
-
-        return combo_list
-
-    def _flush(self):
-        """ A dirty hack to flush the image
-        """
-        w, h = self.get_width_height()
-        self.resize(w+1, h)
-        self.resize(w, h)
-
-    def _setup_legend(self, location='best', font_size=10):
-        """
-        Set up legend
-        self.axes.legend(): Handler is a Line2D object. Lable maps to the line object
-        """
-        allowed_location_list = [
-            "best",
-            "upper right",
-            "upper left",
-            "lower left",
-            "lower right",
-            "right",
-            "center left",
-            "center right",
-            "lower center",
-            "upper center",
-            "center"]
-
-        # Check legend location valid or not
-        if location not in allowed_location_list:
-            location = 'best'
-
-        handles, labels = self.axes.get_legend_handles_labels()
-        self.axes.legend(handles, labels, loc=location, fontsize=font_size)
-
-        self._isLegendOn = True
-
-# END-OF-CLASS (MplGraphicsView)
-
-
-class MyNavigationToolbar(NavigationToolbar2):
-    """ A customized navigation tool bar attached to canvas
-    Note:
-    * home, left, right: will not disable zoom/pan mode
-    * zoom and pan: will turn on/off both's mode
-
-    Other methods
-    * drag_pan(self, event): event handling method for dragging canvas in pan-mode
-    """
-    NAVIGATION_MODE_NONE = 0
-    NAVIGATION_MODE_PAN = 1
-    NAVIGATION_MODE_ZOOM = 2
-
-    # This defines a signal called 'home_button_pressed' that takes 1 boolean
-    # argument for being in zoomed state or not
-    home_button_pressed = Signal()
-
-    # This defines a signal called 'canvas_zoom_released'
-    canvas_zoom_released = Signal()
-
-    def __init__(self, parent, canvas):
-        """ Initialization
-        built-in methods
-        - drag_zoom(self, event): triggered during holding the mouse and moving
-        """
-        NavigationToolbar2.__init__(self, canvas, canvas)
-
-        # parent
-        self._myParent = parent
-        # tool bar mode
-        self._myMode = MyNavigationToolbar.NAVIGATION_MODE_NONE
-
-        # connect the events to parent
-        self.home_button_pressed.connect(self._myParent.evt_toolbar_home)
-        self.canvas_zoom_released.connect(self._myParent.evt_zoom_released)
-
-    @property
-    def is_zoom_mode(self):
-        """
-        check whether the tool bar is in zoom mode
-        """
-        return self._myMode == MyNavigationToolbar.NAVIGATION_MODE_ZOOM
-
-    def get_mode(self):
-        """
-        :return: integer as none/pan/zoom mode
-        """
-        return self._myMode
-
-    # Overriding base's methods
-    def draw(self):
-        """
-        Canvas is drawn called by pan(), zoom()
-        :return:
-        """
-        NavigationToolbar2.draw(self)
-
-        self._myParent.evt_view_updated()
-
-    def home(self, *args):
-        # call super's home() method
-        NavigationToolbar2.home(self, args)
-
-        # send a signal to parent class for further operation
-        self.home_button_pressed.emit()
-
-    def pan(self, *args):
-        """
-
-        :param args:
-        :return:
-        """
-        NavigationToolbar2.pan(self, args)
-
-        if self._myMode == MyNavigationToolbar.NAVIGATION_MODE_PAN:
-            # out of pan mode
-            self._myMode = MyNavigationToolbar.NAVIGATION_MODE_NONE
-        else:
-            # into pan mode
-            self._myMode = MyNavigationToolbar.NAVIGATION_MODE_PAN
-
-        print('PANNED')
-
-    def zoom(self, *args):
-        """
-        Turn on/off zoom (zoom button)
-        :param args:
-        :return:
-        """
-        NavigationToolbar2.zoom(self, args)
-
-        if self._myMode == MyNavigationToolbar.NAVIGATION_MODE_ZOOM:
-            # out of zoom mode
-            self._myMode = MyNavigationToolbar.NAVIGATION_MODE_NONE
-        else:
-            # into zoom mode
-            self._myMode = MyNavigationToolbar.NAVIGATION_MODE_ZOOM
-
-    def release_zoom(self, event):
-        """
-        override zoom released method
-        """
-        self.canvas_zoom_released.emit()
-
-        NavigationToolbar2.release_zoom(self, event)
-
-    def _update_view(self):
-        """
-        view update called by home(), back() and forward()
-        :return:
-        """
-        NavigationToolbar2._update_view(self)
-
-        self._myParent.evt_view_updated()
